feat: Add support for PSR-17/PSR-18 HTTP clients and factories (#2)
* feat: Add support for PSR-17/PSR-18 HTTP clients and factories

This commit adds support for any PSR-17 and PSR-18 compatible HTTP client and factory, removing the previous dependency on Guzzle. The OpenAI class has undergone a major refactoring to support this change, as well as to remove the singleton pattern. Additionally, all examples have been updated to use the updated OpenAI class.

Several optimizations have also been made in this commit, including the refactoring and optimization of test cases in OpenAITest and the OpenAIException class. Finally, the README.md has been updated to reflect changes in the project structure and requirements.